// src/typings/bedrock/schemas/blocks/blocks.d.ts
export type FormatVersion = string;
// ...
export type Permutations = Permutation[];

/**
 * The minecraft block behavior specification.
 */
export interface BlockBehavior {
  format_version: FormatVersion;
  "minecraft:block": BlockDefinitions;
}
/**
 * A custom block definition.
 */
export interface BlockDefinitions {
  description: BlockDescription;
  events?: Events;
  components: Component;
  permutations?: Permutations;
}
/**
 * The description for this block.
 */
export interface BlockDescription {
  menu_category?: MenuCategory;
  identifier: Identifier;
  is_experimental?: IsExperimental;
  register_to_creative_menu?: RegisterToCreativeMenu;
  states?: States;
}
/**
 * [HOLIDAY EXPERIMENTAL] UNDOCUMENTED.
 */
export interface States {
  [k: string]: State;
}
/**
 * UNDOCUMENTED.
 */
export interface Events {
  "minecraft:on_fall_on"?: Event;
  "minecraft:on_interact"?: Event;
  "minecraft:on_placed"?: Event;
  "minecraft:on_player_destroyed"?: Event;
  "minecraft:on_player_placing"?: Event;
  "minecraft:on_step_off"?: Event;
  "minecraft:on_step_on"?: Event;
  [k: string]: unknown;
}
/**
 * UNDOCUMENTED.
 */
export interface Event {
  condition?: Condition;
  event?: Event1;
  target?: Target;
}
/**
 * UNDOCUMENTED.
 */
export interface Component {
  "minecraft:collision_box"?: CollisionBox;
  "minecraft:crafting_table"?: CraftingTable;
  "minecraft:destructible_by_explosion"?: DestructibleByExplosion;
  "minecraft:destructible_by_mining"?: DestructibleByMining;
  "minecraft:display_name"?: DisplayName;
  "minecraft:flammable"?: Flammable;
  "minecraft:friction"?: Friction;
  "minecraft:geometry"?: Geometry;
  "minecraft:light_dampening"?: LightDampening;
  "minecraft:light_emission"?: LightEmission;
  "minecraft:loot"?: Loot;
  "minecraft:map_color"?: MapColor;
  "minecraft:material_instances"?: MaterialInstances;
  "minecraft:part_visibility"?: PartVisibility;
  "minecraft:placement_filter"?: PlacementFilter;
  "minecraft:selection_box"?: SelectionBox;
  "minecraft:transformation"?: Transformation;
  "minecraft:unit_cube"?: UnitCube;
  "minecraft:on_fall_on"?: OnFallOn;
  "minecraft:on_interact"?: OnFallOn1;
  "minecraft:on_placed"?: OnPlaced;
  "minecraft:on_player_destroyed"?: OnPlayerDestroyed;
  "minecraft:on_player_placing"?: OnPlayerPlacing;
  "minecraft:on_step_off"?: OnStepOff;
  "minecraft:on_step_on"?: OnStepOn;
  "minecraft:queued_ticking"?: RandomTicking;
  "minecraft:random_ticking"?: RandomTicking1;
}